# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Описание взаимодействия и работы внутренних моделей E2E CRM",
        "title": "E2E",
        "contact": {
            "name": "Repo",
            "url": "https://github.com/GeekchanskiY/E2E"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/GeekchanskiY/E2E/blob/main/LICENSE"
        },
        "version": "0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/users/register": {
            "post": {
                "description": "register user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "user id",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/finworker_internal_models_requests_users.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users/{userId}": {
            "get": {
                "description": "get user by user id",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "finworker_internal_models_requests_users.RegisterRequest": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}